fix(gallery): reduce file size
