+++ src/lib/viewers/gallery/VirtualizedGalleryGrid.tsx
@@ -9,23 +9,8 @@ import React, {
 } from 'react';
 import { defaultRangeExtractor, useVirtualizer } from '@tanstack/react-virtual';
 import noop from 'lodash/noop';
-import throttle from 'lodash/throttle';
 import { decodeKeydown } from '../../util';
-import HighResThumbnailStore from './HighResThumbnailStore';
-import {
-    CONCURRENT_LOADS,
-    GALLERY_GRID_PADDING_TOP,
-    GALLERY_HIGH_RES_CONCURRENCY,
-    GALLERY_HIGH_RES_MAX_BYTES,
-    GALLERY_HIGH_RES_MAX_PAGES,
-    GALLERY_THUMB_MAX_DPR,
-    GALLERY_THUMB_MAX_TIER,
-    GALLERY_THUMB_MAX_WIDTH,
-    GALLERY_THUMB_WIDTH_TIERS,
-    GALLERY_TILE_GAP,
-    GALLERY_VIRTUAL_OVERSCAN,
-    SCROLL_THROTTLE_MS,
-} from './constants';
+import { GALLERY_GRID_PADDING_TOP, GALLERY_TILE_GAP, GALLERY_VIRTUAL_OVERSCAN } from './constants';
 import {
     getGalleryLayout,
     getOccupiedColumns,
@@ -38,6 +23,7 @@ import {
 import { getColumnIndex, getPageAbove, getPageBelow, getRowCount, getRowIndex } from './galleryGridNavigation';
 import { GalleryTile, GalleryThumbnail } from './GalleryGridShared';
 import useGalleryPinch, { PinchDirection, PinchFocal } from './useGalleryPinch';
+import useGalleryThumbnailLoader from './useGalleryThumbnailLoader';
 import './GalleryGrid.scss';
 
 export type VirtualizedGalleryGridHandle = {
@@ -227,30 +213,19 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
         },
         ref,
     ) {
-        const [loadedImages, setLoadedImages] = useState<Record<number, string>>({});
-        const [highResImages, setHighResImages] = useState<Record<number, string>>({});
         const [focusedPage, setFocusedPage] = useState(currentPage);
-        const [pageRatio, setPageRatio] = useState<number | null>(null);
         const [layoutWidth, setLayoutWidth] = useState(0);
         const anchorPageRef = useRef(currentPage);
         const gridRef = useRef<HTMLDivElement>(null);
         const innerRef = useRef<HTMLDivElement>(null);
         const focalRef = useRef<PinchFocal | null>(null);
         const scaleRef = useRef(scale);
-        const queueRef = useRef<number[]>([]);
-        const isProcessingRef = useRef(false);
-        const isMountedRef = useRef(true);
-        const inFlightRef = useRef<Set<number>>(new Set());
-        const isQueueReadyRef = useRef(false);
-        const highResStoreRef = useRef<HighResThumbnailStore | null>(null);
         const pendingFocusRef = useRef<PendingFocus | null>(null);
-        const loadedImagesRef = useRef(loadedImages);
         const focusedPageRef = useRef(focusedPage);
         const layoutWidthRef = useRef(layoutWidth);
         const pageCountRef = useRef(pageCount);
         const currentPageRef = useRef(currentPage);
 
-        loadedImagesRef.current = loadedImages;
         focusedPageRef.current = focusedPage;
         layoutWidthRef.current = layoutWidth;
         pageCountRef.current = pageCount;
@@ -264,21 +239,19 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
         columnsRef.current = columns;
         tileWidthRef.current = tileWidth;
 
-        const getRatio = useCallback((pageNum: number): number => resolvePageRatio(pageNum, getPageRatio, pageRatio), [
-            getPageRatio,
-            pageRatio,
-        ]);
-        const getRatioRef = useRef(getRatio);
-        getRatioRef.current = getRatio;
-
         const prevLayoutRef = useRef({ columns, tileWidth, scale });
         const prevRatiosRef = useRef<number[] | undefined>(undefined);
         const hasMeasuredLayoutRef = useRef(false);
 
+        // Virtualizer needs getRatioRef for estimateSize; the ref is updated after the
+        // thumbnail loader provides the first-page ratio.
+        const getRatioRef = useRef((pageNum: number): number => resolvePageRatio(pageNum, getPageRatio, null));
+
         const virtualizer = useVirtualizer({
             count: getRowCount(pageCount, columns),
             enabled: true,
-            estimateSize: (rowIndex: number) => getRowHeight(rowIndex, pageCount, columns, tileWidth, getRatio),
+            estimateSize: (rowIndex: number) =>
+                getRowHeight(rowIndex, pageCount, columns, tileWidth, getRatioRef.current),
             gap: GALLERY_TILE_GAP,
             getScrollElement: () => gridRef.current,
             overscan: GALLERY_VIRTUAL_OVERSCAN,
@@ -296,155 +269,27 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
         const virtualizerRef = useRef(virtualizer);
         virtualizerRef.current = virtualizer;
 
-        const byDistanceFromAnchor = (a: number, b: number): number =>
-            Math.abs(a - anchorPageRef.current) - Math.abs(b - anchorPageRef.current);
+        const {
+            loadedImages,
+            highResImages,
+            pageRatio,
+            isQueueReadyRef,
+            handleScrollThrottled,
+        } = useGalleryThumbnailLoader(thumbnail, currentPage, pageCount, {
+            gridRef,
+            virtualizerRef,
+            columnsRef,
+            tileWidthRef,
+            pageCountRef,
+            anchorPageRef,
+            scaleRef,
+        });
 
-        function getNeededThumbWidth(): number {
-            if (scaleRef.current === 1) {
-                return GALLERY_THUMB_MAX_WIDTH;
-            }
-            const neededWidth = tileWidthRef.current * Math.min(window.devicePixelRatio || 1, GALLERY_THUMB_MAX_DPR);
-            return GALLERY_THUMB_WIDTH_TIERS.find(tier => tier >= neededWidth) || GALLERY_THUMB_MAX_TIER;
-        }
-
-        function getPagesNearViewport(marginRatio: number, isEligible?: (pageNum: number) => boolean): number[] {
-            const grid = gridRef.current;
-            if (!grid) return [];
-
-            const { scrollTop, clientHeight } = grid;
-            if (clientHeight <= 0) {
-                return [];
-            }
-
-            const viewportBottom = scrollTop + clientHeight;
-            const margin = clientHeight * marginRatio;
-            const visible: number[] = [];
-            const nearby: number[] = [];
-            const columnCount = columnsRef.current;
-            const pages = pageCountRef.current;
-
-            virtualizerRef.current.getVirtualItems().forEach(item => {
-                const bottom = item.start + item.size;
-                let bucket: number[] | null = null;
-                if (bottom > scrollTop && item.start < viewportBottom) {
-                    bucket = visible;
-                } else if (bottom > scrollTop - margin && item.start < viewportBottom + margin) {
-                    bucket = nearby;
-                }
-                if (!bucket) {
-                    return;
-                }
-                const target = bucket;
-                getPagesInRow(item.index, columnCount, pages).forEach(pageNum => {
-                    if (!isEligible || isEligible(pageNum)) {
-                        target.push(pageNum);
-                    }
-                });
-            });
-
-            return [...visible.sort(byDistanceFromAnchor), ...nearby.sort(byDistanceFromAnchor)];
-        }
-
-        function syncHighRes() {
-            const store = highResStoreRef.current;
-            const { pageRatio: thumbRatio } = thumbnail;
-            if (!store || !thumbRatio) return;
-
-            const width = getNeededThumbWidth();
-            if (width === GALLERY_THUMB_MAX_WIDTH) {
-                store.setRetained([], width, thumbRatio);
-            } else if (!isProcessingRef.current) {
-                store.setRetained(getPagesNearViewport(0.5), width, thumbRatio);
-            }
-        }
-
-        function getUnloadedNearViewport(): number[] {
-            return getPagesNearViewport(3, (pageNum: number) => {
-                if (inFlightRef.current.has(pageNum)) {
-                    return false;
-                }
-                return !loadedImagesRef.current[pageNum];
-            });
-        }
-
-        function processQueue() {
-            if (!isMountedRef.current || !thumbnail || queueRef.current.length === 0) {
-                isProcessingRef.current = false;
-                return;
-            }
-
-            const batch = queueRef.current.slice(0, CONCURRENT_LOADS);
-            queueRef.current = queueRef.current.slice(CONCURRENT_LOADS);
-
-            requestAnimationFrame(() => {
-                if (!isMountedRef.current || !thumbnail) {
-                    isProcessingRef.current = false;
-                    return;
-                }
-
-                let completed = 0;
-
-                const onComplete = () => {
-                    completed += 1;
-                    if (completed < batch.length) return;
-
-                    if (!isMountedRef.current) {
-                        isProcessingRef.current = false;
-                        return;
-                    }
-
-                    const remaining = new Set(queueRef.current);
-                    queueRef.current = getUnloadedNearViewport().filter(p => remaining.has(p));
-                    if (queueRef.current.length === 0) {
-                        isProcessingRef.current = false;
-                        syncHighRes();
-                        return;
-                    }
-                    processQueue();
-                };
-
-                batch.forEach(pageNum => {
-                    inFlightRef.current.add(pageNum);
-                    thumbnail
-                        .createThumbnailImage(pageNum - 1, {
-                            createImgTag: true,
-                            thumbMaxWidth: GALLERY_THUMB_MAX_WIDTH,
-                        })
-                        .then((imageEl: HTMLImageElement | null) => {
-                            inFlightRef.current.delete(pageNum);
-                            if (isMountedRef.current && imageEl && imageEl.src) {
-                                setLoadedImages(prev => ({ ...prev, [pageNum]: imageEl.src }));
-                            }
-                            onComplete();
-                        })
-                        .catch(err => {
-                            inFlightRef.current.delete(pageNum);
-                            console.error('Gallery thumbnail failed for page', pageNum, err); // eslint-disable-line no-console
-                            onComplete();
-                        });
-                });
-            });
-        }
-
-        function startProcessing() {
-            if (!isProcessingRef.current && queueRef.current.length > 0) {
-                isProcessingRef.current = true;
-                processQueue();
-            }
-        }
-
-        const handleScrollRef = useRef(
-            throttle(() => {
-                const nearbyUnloaded = getUnloadedNearViewport();
-                if (nearbyUnloaded.length > 0) {
-                    const currentQueue = new Set(queueRef.current);
-                    const toAdd = nearbyUnloaded.filter(p => !currentQueue.has(p));
-                    queueRef.current = [...toAdd, ...queueRef.current];
-                    startProcessing();
-                }
-                syncHighRes();
-            }, SCROLL_THROTTLE_MS),
-        );
+        const getRatio = useCallback((pageNum: number): number => resolvePageRatio(pageNum, getPageRatio, pageRatio), [
+            getPageRatio,
+            pageRatio,
+        ]);
+        getRatioRef.current = getRatio;
 
         const handleScroll = useCallback(() => {
             const grid = gridRef.current;
@@ -457,8 +302,8 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
                     }
                 }
             }
-            handleScrollRef.current();
-        }, []);
+            handleScrollThrottled();
+        }, [handleScrollThrottled]);
 
         useGalleryPinch({
             focalRef,
@@ -470,6 +315,7 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
             scaleRef,
         });
 
+        // Remeasure row heights when page ratios change.
         // eslint-disable-next-line react-hooks/exhaustive-deps
         useLayoutEffect(() => {
             if (!hasMeasuredLayoutRef.current || layoutWidth <= 0) {
@@ -517,7 +363,7 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
                 grid.scrollTop += nextStart - prevStart;
             }
             prevRatiosRef.current = collectPageRatios(pageCount, getRatio);
-            handleScrollRef.current();
+            handleScrollThrottled();
         });
 
         useLayoutEffect(() => {
@@ -527,6 +373,7 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
             }
         }, [columns, getRatio, pageCount, tileWidth, virtualizer]);
 
+        // Restore scroll position after scale/column/width changes.
         useLayoutEffect(() => {
             const grid = gridRef.current;
             const previousScale = scaleRef.current;
@@ -556,7 +403,7 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
                     prevRatiosRef,
                     pendingFocusRef,
                     scrollToIndex: virtualizer.scrollToIndex,
-                    onReady: handleScrollRef.current,
+                    onReady: handleScrollThrottled,
                 })
             ) {
                 return;
@@ -576,9 +423,20 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
                 tileWidth,
                 getRatio,
             );
-            handleScrollRef.current();
-        }, [columns, currentPage, getRatio, layoutWidth, pageCount, scale, tileWidth, virtualizer]);
+            handleScrollThrottled();
+        }, [
+            columns,
+            currentPage,
+            getRatio,
+            handleScrollThrottled,
+            layoutWidth,
+            pageCount,
+            scale,
+            tileWidth,
+            virtualizer,
+        ]);
 
+        // Apply pending focus after virtualizer commits new rows.
         useLayoutEffect(() => {
             const pending = pendingFocusRef.current;
             if (!pending) {
@@ -596,73 +454,13 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
             }
         });
 
+        // Set initial pending focus on mount.
         useEffect(() => {
-            const throttledScroll = handleScrollRef.current;
-            const highResStore = new HighResThumbnailStore({
-                maxBytes: GALLERY_HIGH_RES_MAX_BYTES,
-                maxConcurrent: GALLERY_HIGH_RES_CONCURRENCY,
-                maxPages: GALLERY_HIGH_RES_MAX_PAGES,
-                onChange: images => {
-                    if (isMountedRef.current) {
-                        setHighResImages(images);
-                    }
-                },
-                render: (pageNum, width) => thumbnail.renderPageImage(pageNum, { thumbMaxWidth: width }),
-            });
-            highResStoreRef.current = highResStore;
-
             pendingFocusRef.current = pendingFocusRef.current ?? { page: currentPage, from: document.activeElement };
-
-            const initialImages: Record<number, string> = {};
-            const uncachedPages: number[] = [];
-
-            for (let i = 1; i <= pageCount; i += 1) {
-                const cached = thumbnail.getImageFromCache(i - 1);
-                const src = cached?.image?.src;
-                if (src) {
-                    initialImages[i] = src;
-                } else {
-                    uncachedPages.push(i);
-                }
-            }
-
-            if (Object.keys(initialImages).length > 0) {
-                setLoadedImages(initialImages);
-            }
-
-            uncachedPages.sort((a, b) => Math.abs(a - currentPage) - Math.abs(b - currentPage));
-            queueRef.current = uncachedPages;
-            isQueueReadyRef.current = true;
-
-            thumbnail
-                .init()
-                .then(() => {
-                    if (!isMountedRef.current) return;
-                    if (typeof thumbnail.pageRatio === 'number' && thumbnail.pageRatio > 0) {
-                        setPageRatio(thumbnail.pageRatio);
-                    }
-                    startProcessing();
-                    syncHighRes();
-                })
-                .catch(err => {
-                    if (!isMountedRef.current) return;
-                    console.error('Gallery thumbnail init failed', err); // eslint-disable-line no-console
-                    startProcessing();
-                });
-
-            return () => {
-                isMountedRef.current = false;
-                isQueueReadyRef.current = false;
-                pendingFocusRef.current = null;
-                queueRef.current = [];
-                isProcessingRef.current = false;
-                highResStore.destroy();
-                highResStoreRef.current = null;
-                throttledScroll.cancel();
-            };
             // eslint-disable-next-line react-hooks/exhaustive-deps
         }, []);
 
+        // ResizeObserver for initial width measurement and reflow.
         useEffect(() => {
             const grid = gridRef.current;
             const inner = innerRef.current;
@@ -692,18 +490,19 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
                         prevRatiosRef,
                         pendingFocusRef,
                         scrollToIndex: virtualizerRef.current.scrollToIndex,
-                        onReady: handleScrollRef.current,
+                        onReady: handleScrollThrottled,
                     });
                 }
                 if (isFirstObservation) {
                     isFirstObservation = false;
                     return;
                 }
-                handleScrollRef.current();
+                handleScrollThrottled();
             });
             observer.observe(grid);
 
             return () => observer.disconnect();
+            // eslint-disable-next-line react-hooks/exhaustive-deps
         }, []);
 
         const selectPage = useCallback(
@@ -850,6 +649,7 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
 
         const virtualItems = virtualizer.getVirtualItems();
 
+        // Pump thumbnail loading when focus moves to ensure new rows get thumbnails.
         const skipInitialFocusPumpRef = useRef(true);
         useEffect(() => {
             if (!isQueueReadyRef.current) {
@@ -861,12 +661,12 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
             }
             const pump = (): void => {
                 handleScroll();
-                handleScrollRef.current.flush();
+                handleScrollThrottled.flush();
             };
             pump();
             const raf = requestAnimationFrame(pump);
             return () => cancelAnimationFrame(raf);
-        }, [focusedPage, handleScroll]);
+        }, [focusedPage, handleScroll, handleScrollThrottled, isQueueReadyRef]);
 
         const renderTile = (pageNum: number): JSX.Element => (
             <GalleryTile
@@ -891,6 +691,8 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
             const pages = getPagesInRow(virtualRow.index, columns, pageCount);
 
             return (
+                // ARIA grid, not an HTML table — <tr> is not valid here.
+                // eslint-disable-next-line jsx-a11y/prefer-tag-over-role
                 <div
                     key={virtualRow.key}
                     aria-label={String(rowIndex)}
@@ -922,12 +724,7 @@ const VirtualizedGalleryGrid = forwardRef<VirtualizedGalleryGridHandle, Virtuali
                 role="grid"
                 tabIndex={-1}
             >
-                <div
-                    ref={innerRef}
-                    className="bp-gallery-grid-inner"
-                    role="presentation"
-                    style={{ height: virtualizer.getTotalSize() }}
-                >
+                <div ref={innerRef} className="bp-gallery-grid-inner" style={{ height: virtualizer.getTotalSize() }}>
                     {content}
                 </div>
             </div>
